demo data were added to commandliner, in book controller method to heandle the get request were added which fetch all data from DB and shows it in booklist.html, in the Book entity class the price attribute were change to duble type,

+++ bookstore/src/main/java/com/plakolli/bookstore/domain/Book.java
@@ -13,13 +13,13 @@ public class Book {
 	private Long id;
 	private String title, author, isbn;
 	private int year;
-	private Float price;
+	private double price;
 	
 	
 	
 	public Book() {} // constructor without parameters.
 
-	public Book(String title, String author, String isbn, int year, Float price) {
+	public Book(String title, String author, String isbn, int year, double price) {
 		this.title = title;
 		this.author = author;
 		this.isbn = isbn;
@@ -67,7 +67,7 @@ public class Book {
 		this.year = year;
 	}
 
-	public Float getPrice() {
+	public double getPrice() {
 		return price;
 	}
